requirements.txt
a2wsgi==1.10.7
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
bentoml==1.3.13
cattrs==23.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
deepmerge==2.0
defusedxml==0.7.1
deprecated==1.2.15
filelock==3.13.1
fonttools==4.55.0
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-ws==0.6.2
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
jinja2==3.1.3
kiwisolver==1.4.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.3
nvidia-ml-py==12.560.30
opencv-python==4.10.0.84
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.2.0
pip-requirements-parser==32.0.1
prometheus-client==0.21.0
prompt-toolkit==3.0.36
propcache==0.2.0
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.17
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
questionary==2.0.1
requests==2.32.3
rich==13.9.4
schema==0.7.7
scipy==1.14.1
seaborn==0.13.2
setuptools==70.0.0
simple-di==0.1.5
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
supervision==0.25.0
sympy==1.13.1
tomli-w==1.1.0
torch==2.5.1+cpu
torchvision==0.20.1+cpu
tornado==6.4.1
tqdm==4.67.0
typing-extensions==4.9.0
tzdata==2024.2
ultralytics==8.3.34
ultralytics-thop==2.0.12
urllib3==2.2.3
uv==0.5.3
uvicorn==0.32.0
watchfiles==0.24.0
wcwidth==0.2.13
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.2
zipp==3.21.0
